Split the particle renderer into lifecycle, children and snapshot files
